+++ frontend/src/components/BetsAndResuls/DailyScore.tsx
