+++ frontend/src/components/BetsAndResuls/GameResult.tsx
